requirements.txt
aenum==3.1.12
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.26.151
botocore==1.29.151
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.35
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
Fiona==1.9.4.post1
fonttools==4.39.4
GDAL==3.4.3
geojson==3.0.1
geomet==1.0.0
geopandas==0.13.2
html2text==2020.1.16
idna==3.4
imageio==2.31.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.6.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
kiwisolver==1.4.4
lazy-loader==0.2
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook-shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.3
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.19.3
pysolar==0.10
pytest==7.3.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
s3transfer==0.6.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
sentinelhub==3.9.1
sentinelsat==1.2.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
tomli-w==1.0.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.16
utm==0.7.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.0
widgetsnbextension==3.6.4
zipp==3.15.0
